+++ src/entry/Entry.tsx
@@ -2,7 +2,7 @@ import React, { Fragment } from 'react';
 import { withRouter, RouteComponentProps } from 'react-router-dom';
 import useDataApi from '../utils/useDataApi';
 import apiUrls from '../utils/apiUrls';
-import FieldToViewMappings from '../views/FieldToViewMappings';
+import { EntryProteinNames } from '../model/ProteinNames';
 
 interface MatchParams {
   accession: string;
@@ -17,10 +17,11 @@ const Entry: React.FC<EntryProps> = ({ match }) => {
     return null;
   }
   const entryData = data.results[0];
+
   return (
     <Fragment>
-      {FieldToViewMappings.protein_name(entryData)}
-      {FieldToViewMappings.organism(entryData)}
+      <h3>Names &amp; Taxonomy</h3>
+      <EntryProteinNames data={entryData} />
     </Fragment>
   );
 };
