+++ src/model/NameView.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 type NameViewProps = {
   name: string | null | undefined;
   shortName?: string | null | undefined;
-  alternativeNames?: Array<string>;
+  alternativeNames?: string[];
 };
 
 const NameView = (props: NameViewProps) => {
@@ -11,18 +11,12 @@ const NameView = (props: NameViewProps) => {
   let altNames;
   if (alternativeNames && alternativeNames.length > 0) {
     altNames = alternativeNames.map(alternativeName => (
-      <span key={alternativeName}>
-        {' '}
-        {alternativeName}
-      </span>
+      <span key={alternativeName}> {alternativeName}</span>
     ));
   }
   return (
     <div>
-      <strong>{name}</strong>
-      {' '}
-      {shortName}
-      {' '}
+      <strong>{name}</strong> {shortName}{' '}
       {altNames ? <span>{altNames}</span> : undefined}
     </div>
   );
