+++ src/model/SimpleView.tsx
