--- src/views/__tests__/FieldToViewMappings.spec.tsx
@@ -1,81 +1,0 @@
-import { shallow, configure } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
-import FieldToViewMappings from '../FieldToViewMappings';
-
-configure({ adapter: new Adapter() });
-
-describe('FieldToViewMappings', () => {
-  test('should render protein_name', () => {
-    const data = {
-      proteinDescription: {
-        recommendedName: {
-          fullName: {
-            value: 'My protein',
-          },
-        },
-        submissionNames: [
-          {
-            fullName: {
-              value: 'A submission',
-            },
-          },
-        ],
-        shortNames: [
-          {
-            value: 'My prot',
-          },
-        ],
-        ecNumbers: [
-          {
-            value: 'ec1234',
-          },
-        ],
-        alternativeNames: [
-          {
-            fullName: {
-              value: 'Some protein',
-            },
-          },
-        ],
-      },
-    };
-    const component = FieldToViewMappings.protein_name(data);
-    const wrapper = shallow(component);
-    expect(wrapper.debug()).toMatchSnapshot();
-  });
-
-  test('should render gene_names', () => {
-    const data = {
-      genes: [
-        {
-          geneName: {
-            value: 'My gene',
-          },
-          synonyms: [{ value: 'synonym 1' }, { value: 'synonym 2' }],
-          orfNames: [
-            { value: 'orfNames 1' },
-            { value: 'orfNames 2' },
-            { value: 'orfNames 3' },
-          ],
-        },
-      ],
-    };
-    const component = FieldToViewMappings.gene_names(data);
-    const wrapper = shallow(component);
-    expect(wrapper.debug()).toMatchSnapshot();
-  });
-
-  test('should render organism', () => {
-    const data = {
-      organism: {
-        scientificName: 'Dodo scientifacus',
-        commonName: 'Lesser spotted dodo',
-        synonyms: ['Flightless bird', "Darwin's pet"],
-        taxonId: 1234,
-      },
-    };
-    const component = FieldToViewMappings.organism(data);
-    const wrapper = shallow(component);
-    expect(wrapper.debug()).toMatchSnapshot();
-  });
-});
