+++ tslint.json
@@ -13,7 +13,7 @@
     "newline-before-return": false,
     "no-any": false,
     "no-empty-interface": false,
-    "no-import-side-effect": [true],
+    "no-import-side-effect": [true, { "ignore-pattern": "*.css" }],
     "no-inferrable-types": [true, "ignore-params", "ignore-properties"],
     "no-invalid-this": [true, "check-function-in-method"],
     "no-namespace": false,
